# PERF: Make ImageGridSampler multi-threaded